mp-buildbot: Accept multi-inputs for list-subports
For the use case in our buildbot setup, we would prefer calling ./mpbb
list-subports only once with a list of all ports for which we want their
subports, so add this possibility using positional arguments.

Additionally, filter port names that are not found in the portindex to avoid
scheduling useless builds on the port builders.

git-svn-id: https://svn.macports.org/repository/macports/contrib/mp-buildbot@151029 d073be05-634f-4543-b044-5fe20cf6d1d6