mp-buildbot: Install dependencies one-by-one
Install dependencies step by step in install-dependencies so that we can track
which dependencies failed and start building a fail cache again. Additionally,
this fixes a problem where
 > port echo depof:wine
would print a list of dependencies but those dependencies were not installed
with the +universal variant as required. To do this, tools/dependencies.tcl
calls mportdepends, which determines the necessary variants of the
dependencies.

git-svn-id: https://svn.macports.org/repository/macports/contrib/mp-buildbot@151047 d073be05-634f-4543-b044-5fe20cf6d1d6