mp-buildbot: Support variants in dependency calc
Adding a failcache to avoid wasting time requires a path per port that
we can use to cache the result of the last build. Since we're planning
to support variants in mp-buildbot, this path needs to be
variant-specific as well. In order to implement the fail cache, we thus
need working variant support first.

Fix a bug in the variant calculation of dependencies that would trigger
in the following situation:

 port bar:
   depends_lib port:foo

 port foo:
   default_variants +foo +bar

 ./tools/dependencies.tcl bar -bar

This setup used to print "foo +foo" only, but passing that to "port
install" would still install foo with the bar variant. With this change,
the correct value "foo -bar+foo" is printed instead.

git-svn-id: https://svn.macports.org/repository/macports/contrib/mp-buildbot@152489 d073be05-634f-4543-b044-5fe20cf6d1d6