mpbb: Parse command-line arguments with a function
This will facilitate the separation of global options and subcommand
options because each subcommand function will be able to parse its own
arguments with only a few lines of code instead of an unwieldy `case`
statement.

The new function requires enhanced getopt(1) [*], which is not included
with macOS but is available in the `getopt` port. It recognizes flags in
`--OPT=ARG` form, permits optional arguments, and correctly handles
missing arguments.

[*]: http://frodo.looijaard.name/project/getopt

git-svn-id: https://svn.macports.org/repository/macports/contrib/mp-buildbot@152524 d073be05-634f-4543-b044-5fe20cf6d1d6