mp-buildbot: Don't re-run previously failed builds
If a port's build has previously failed, we do not want to waste cycles
to re-attempt it (e.g. as dependencies of other ports), unless the
failed port's definition has changed.

To achieve that, create a file with a reference to the last failed build
for every failing port, and fail builds that depend on failed ports
early.

git-svn-id: https://svn.macports.org/repository/macports/contrib/mp-buildbot@152581 d073be05-634f-4543-b044-5fe20cf6d1d6