add successcache support to mpbb
Check if the archive for the requested port exists in its installed
location, and if so, simply skip installing.

Do the same for each dependency. Dependencies are still built if not
already installed (even if the requested port is somehow already
installed), so their archives can be uploaded.

Deactivate all ports between builds when installing dependencies.
Activate all the dependencies after they are all installed iff the
requested port will need to be built.

Fixes: https://trac.macports.org/ticket/52765