Do not request default variants explictly
This avoids the problem that explicitly requested variants are passed
down to dependencies, where they can cause conflicts.

This may lead to cache misses in the failcache, as we can no longer
differentiate between a build with an explicitly requested variant and
the default variant. These cases should be rare, so they are negligible.

See: https://lists.macports.org/pipermail/macports-dev/2017-June/035978.html