Set unrequested flag when installing ports
We already install dependencies with the unrequested flag. This now also
installs the main port that way. There seems to be no reason to make a
distinction between the main port and its dependencies with regard to
this flag, since the idea is for all of the ports (or at least, all of
the ports that are dependencies of another port) to remain installed,
regardless of whether they were originally installed explicitly or as a dependency.

Closes: https://trac.macports.org/ticket/57571