Do not restrict which builders build noarch ports
The arm64 Buildbot builder is faster so we want to build most noarch
ports there. Similarly, the 10.6 i386 builder is idle more often so we
want to build ports there when we can. But some noarch ports or their
dependencies might only be buildable on the x86_64 builders. We also
want ports to be built on the GitHub Actions CI infrastructure and right
now those are x86_64 only.

In the situation where both Buildbot builders are available at the same
time and both try to build it, I assume that buildbot master is
single-threaded and will only process one builder's upload at a time so
that only one builder's archive and its signature will end up on the
package server. Which one is undefined and I assume that is not
important.

Closes: https://trac.macports.org/ticket/62977