mpbb-cleanup: only uninstall ports in watcher builds
It's reasonably common for ports to have a number of dependencies that
aren't used by anything else. Installing one of these ports for the
first time performs very poorly when ports with one or less dependents
are uninstalled between every build, as many of the deps that are
built are immediately uninstalled, only to be installed (and
uninstalled) again and again in subsequent builds.

Only uninstalling these ports as part of portwatcher builds should
greatly mitigate this effect.