Allow deleting distfiles if distfiles is a symlink
When I originally prevented deletions if distfiles was a symlink in
f151333 I was considering replacing the distfiles directory on each
worker with a symlink to the master distfiles volume mounted over
AFP/SMB to save disk space on the workers. I didn't end up trying that,
but on the Apple Silicon workers I have now made distfiles a symlink to
a shared directory where I do want deletions to occur, so I've changed
the condition so now deletions can occur unless ports.tar.gz exists;
that would only exist on the master. I added a slash to the end of the
distfiles path since otherwise find won't descend into the symlinks.