dependencies.tcl: attempt to mitigate 503 errors
Ensure that archivefetch is not attempted at all in install_dep_source,
since we only get there if it already failed. This should prevent
failure from archivefetch getting a 503 for a signature again.

This is at best a poor and partial workaround, since archivefetch
failing for ports that have an archive available in the first place
means that those ports are unnecessarily built from source, and
furthermore, fetching the distfiles might fail the same way.

See: https://trac.macports.org/ticket/68563